Hook and modify all existing buttons on enable
DTK QoL +use: When the plugin is enabled, all existing buttons are hooked and their spawnflags modified. This is to counter a problem in TF2 Classic where the plugin is enabled via mapname config after the first set of entities has been created, and joining the game alone does not cause a round restart.