Added option to remove weapons
In restrictions.cfg you can now tell the plugin to remove a weapon instead of replace it or add attributes. This is useful for things like spy watches, when you do not want spies to be able to cloak. Avoid removing primary weapons (or melee weapons if melee-only mode is enabled) as players will go into reference pose until they switch to an alternative.